requirements/requirements.txt
crispy-forms-foundation
Django
django-admin-bootstrapped
django-bootstrap3
django-bower
django-braces
django-compressor
django-crispy-forms
django-extra-views
django-haystack
django-vanilla-views
git+git://github.com/interlegis/django-sass-processor.git
git+git://github.com/LeandroRoberto/odfpy.git
libsass
psycopg2
pytz
PyYAML
trml2pdf>=0.3
Unipath
whoosh
